use ExpressionTypeResolverExtension instead of DynamicMethodReturnTypeExtension
